+++ BetterSplitscreen/Compatibility/IGenericModConfigMenuApi.cs
@@ -5,7 +5,7 @@ using StardewModdingAPI;
 using StardewModdingAPI.Utilities;
 using StardewValley;
 
-namespace GenericModConfigMenu
+namespace BetterSplitscreen.Compatibility
 {
     /// <summary>The API which lets other mods add a config UI through Generic Mod Config Menu.</summary>
     public interface IGenericModConfigMenuApi
