+++ BetterSplitscreen/Layout/LayoutPreviewHelper.cs
@@ -1,11 +1,10 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using StardewValley;
-using System;
 
 namespace BetterSplitscreen.Layout
 {
-    internal class LayoutPreviewHelper
+    internal static class LayoutPreviewHelper
     {
         private static Color P1Color { get; } = new Color(60, 90, 220);
 
@@ -15,17 +14,19 @@ namespace BetterSplitscreen.Layout
 
         private static Color P4Color { get; } = new Color(255, 220, 30);
 
-        public int PlayerCount { get; set; } = 2;
+        public static int PlayerCount { get; set; } = 2;
 
         /// <summary>
         /// Draws a preview of the current layout with the number of players provided by <see cref="PlayerCount"/>.
         /// </summary>
-        /// <param name="sb"></param>
-        /// <param name="p"></param>
-        public void DrawPreview(SpriteBatch sb, Vector2 p)
+        /// <param name="sb">SpriteBatch</param>
+        /// <param name="p">Vector2</param>
+        public static void DrawPreview(SpriteBatch sb, Vector2 p)
         {
             int px = (int)p.X;
             int py = (int)p.Y;
+
+            // TODO: If Mod or Layout Feature disabled, draw "Preview Disabled" text and nothing else.
 
             // Border
             sb.Draw(Game1.fadeToBlackRect, new Rectangle(px, py, 208, 208), Color.Black);
@@ -57,11 +58,6 @@ namespace BetterSplitscreen.Layout
             // TODO: P1, P2, P3, P4; dynamically
             sb.DrawString(Game1.dialogueFont, "P1", new Vector2(p1_x + 30, p1_y + 30), Color.Black);
             sb.DrawString(Game1.dialogueFont, "P2", new Vector2(p2_x + 30, p2_y + 30), Color.Black);
-        }
-
-        internal void OnPreviewChanged(string arg1, object arg2)
-        {
-            throw new NotImplementedException();
         }
     }
 }
